requirements.txt
streamlit
pandas
plotly
openpyxl
python-dotenv
google-genai
